+++ OS/Framework/Timer/Timer.h
@@ -5,7 +5,7 @@
 // Avoid recursive include
 namespace MatrixOS::SYS
 {
-  uint32_t Millis(void);
+  uint64_t Millis(void);
 }
 
 class Timer {
